requirements.txt
requests>=2.31.0
pandas>=2.0.0
thefuzz>=0.20.0
python-Levenshtein>=0.21.0
phonenumbers>=8.13.0
dnspython>=2.4.0
python-dotenv>=1.0.0
